Fixed wobsocket path
Allow path to be passed into websocket connection
so that if there is a proxy in between it can be
setup with a path on the proxy server even though
it doesn't matter for the server connection.

Also there is a future option we can introduce to
pass HTTP headers in case application might use them
to authenticate with the proxy for example.